pyproject.toml
[project]
name = "sft-label"
version = "0.1.0"
description = "SFT Capability Taxonomy & Auto-Labeling Pipeline"
license = "Apache-2.0"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
sft-label = "sft_label.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sft_label"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
